+++ src/Rest/Query/Sort/SortException.php
@@ -8,6 +8,7 @@ class SortException extends \Exception
 {
     public const INVALID_QUERY_PARAMETER = 1;
     public const ATTRIBUTE_PATH_MISSING = 2;
-    public const SORT_KEYS_UNDEFINED = 3;
+    public const SORT_KEY_UNDEFINED = 3;
     public const ATTRIBUTE_PATH_UNDEFINED = 4;
+    public const INVALID_SORT_DIRECTION = 5;
 }